mqtt/bridge/requirements-mqtt.txt
kafka-python==2.0.2
paho-mqtt==1.6.1
python-snappy==0.6.1
